Move StateTransitions to Extension
