+++ src/Extension/CalculableState.php
@@ -4,7 +4,7 @@
 namespace Wirecard\Order\State\Extension;
 
 
-use Wirecard\Order\State\Implementation\StateTransitions;
+use Wirecard\Order\State\Extension\StateTransitions;
 use Wirecard\Order\State\State;
 
 interface CalculableState extends StateTransitions, State
