+++ src/Extension/StateTransitions.php
@@ -1,8 +1,9 @@
 <?php
 
 
-namespace Wirecard\Order\State\Implementation;
+namespace Wirecard\Order\State\Extension;
 
+use Wirecard\Order\State\Implementation\TransitionData;
 use Wirecard\Order\State\State;
 
 interface StateTransitions
